# lib/api-zod/src/generated/types/choice.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * QuizNET API - Plateforme d'entraînement aux QCM
 * OpenAPI spec version: 0.1.0
 */

export interface Choice {
  id: number;
  text: string;
  isCorrect: boolean;
}