# lib/api-zod/src/generated/types/importResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * QuizNET API - Plateforme d'entraînement aux QCM
 * OpenAPI spec version: 0.1.0
 */

export interface ImportResult {
  imported: number;
  errors: string[];
}