# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * QuizNET API - Plateforme d'entraînement aux QCM
 * OpenAPI spec version: 0.1.0
 */
import type { Choice } from './choice';
import type { QuestionType } from './questionType';

export interface Question {
  id: number;
  quizId: number;
  type: QuestionType;
  text: string;
  /** @nullable */
  explanation?: string | null;
  orderIndex: number;
  choices: Choice[];
  /** @nullable */
  directAnswer?: string | null;
}